requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
tzdata==2024.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
black==24.8.0
decorator==5.1.1
ipython==8.26.0
python-crontab==3.2.0
idna==3.7
requests==2.32.3
stripe==10.6.0
urllib3==2.2.2
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
pytz==2024.1
pyyaml==6.0.2
coverage==7.6.1
